link to new docs page
Also clean up lots of dead content. I really would have liked to do this
properly, but the sources are not available. So this is just manually
editing HTML files to get it looking like we want it.

This page can hopefully be replaced with something more usable in the future.

Signed-off-by: Moritz Wanzenböck <moritz.wanzenboeck@linbit.com>